+++ src/phiprof.cpp
@@ -124,27 +124,5 @@ namespace phiprof
    int getChildId(const string &label){
       return parallelTimerTree.getChildId(label);
    }
-
-   Timer::Timer(const int id) : id {id} {
-      this->start();
-   }
-
-   Timer::Timer(const string& label, const vector<string>& groups) : Timer(initializeTimer(label, groups)) {}
-
-   Timer::~Timer() {
-      this->stop();
-   }
-
-   bool Timer::start() {
-      return active ? false : (active = phiprof::start(this->id));
-   }
-
-   bool Timer::stop(const double workUnits, const string& workUnitLabel) {
-      if (!active)
-         return false;
-
-      active = false;
-      return phiprof::stop(id, workUnits, workUnitLabel);
-   }
    
 }
